Remove an obsolete TODO-entry
When trying to test the code for this case, various race-conditions
start appearing. This may be due to the fact, the KVO information is
associated with classes and is being leaked.

Since class objects remain the same between invocations of test methods,
the stop being isolated, and test results become unreliable.